improved custom code handling; improved project problem checking

+++ src/desktop/Awb.Core/Export/SrcFolderCloner.cs
@@ -11,21 +11,27 @@ namespace Awb.Core.Export
 {
     internal class SrcFolderCloner
     {
+        private readonly string _sourceFolder;
+        private readonly string _targetFolder;
+        private readonly bool _removeExtraFilesInTarget;
+        private readonly string[] _removeFilesBlockerDirectoryNames;
+        private readonly bool _deepReportLevel;
+
         public class CloneResult
         {
             public bool Success => string.IsNullOrEmpty(ErrorMessage);
             public string? ErrorMessage { get; set; }
         }
 
-        private readonly string _sourceFolder;
-        private readonly string _targetFolder;
+        public event EventHandler<ExporterProcessStateEventArgs>? Processing;
 
-        public SrcFolderCloner(string sourceFolder, string targetFolder, bool removeExtraFilesInTarget)
+        public SrcFolderCloner(string sourceFolder, string targetFolder, bool removeExtraFilesInTarget, string[] removeFilesBlockerDirectoryNames, bool deepReportLevel=false)
         {
-            if (removeExtraFilesInTarget) throw new System.NotImplementedException("removeExtraFilesInTarget not supported yet!");
-
             _sourceFolder = sourceFolder;
             _targetFolder = targetFolder;
+            _removeExtraFilesInTarget = removeExtraFilesInTarget;
+            _removeFilesBlockerDirectoryNames = removeFilesBlockerDirectoryNames;
+            _deepReportLevel = deepReportLevel;
         }
 
         public async Task<CloneResult> Clone()
@@ -33,15 +39,34 @@ namespace Awb.Core.Export
             if (!Directory.Exists(_sourceFolder)) return new CloneResult { ErrorMessage = $"Source folder '{_sourceFolder}' does not exist" };
             if (!Directory.Exists(_targetFolder)) return new CloneResult { ErrorMessage = $"Target folder '{_targetFolder}' does not exist" };
 
-
-            // Delete files not existing in source folder
-            foreach (var targetFile in Directory.GetFiles(_targetFolder, "*", SearchOption.AllDirectories))
+            // remove files in target folder that are not in source folder
+            if (_removeExtraFilesInTarget)
             {
-                var relativePath = targetFile.Substring(_targetFolder.Length + 1);
-                var sourceFile = Path.Combine(_sourceFolder, relativePath);
+                foreach (var targetFile in Directory.GetFiles(_targetFolder, "*", SearchOption.AllDirectories))
+                {
+                    var relativePath = targetFile.Substring(_targetFolder.Length + 1);
+                    var sourceFile = Path.Combine(_sourceFolder, relativePath);
 
-                if (!File.Exists(sourceFile))
-                    File.Delete(targetFile);
+                    if (!File.Exists(sourceFile))
+                    {
+                        var delete = true;
+                        foreach (var blockerDir in _removeFilesBlockerDirectoryNames)
+                        {
+                            if (sourceFile.Contains($@"\{blockerDir}\"))
+                            {
+                                if (_deepReportLevel == true)
+                                    Processing?.Invoke(this, new ExporterProcessStateEventArgs { Message = $"Skipping deleting file '{targetFile}' because it is located in '{blockerDir}'" });
+                                delete = false;
+                                break;
+                            }
+                        }
+                        if (delete)
+                        {
+                            File.Delete(targetFile);
+                            Processing?.Invoke(this, new ExporterProcessStateEventArgs { Message = $"Deleted file '{targetFile}' because it is not in source folder." });
+                        }
+                    }
+                }
             }
 
             // clone the source folder to the target folder stepping through all files and subfolders
@@ -55,7 +80,8 @@ namespace Awb.Core.Export
 
                 File.Copy(sourceFile, targetFile, true);
             }
-          
+
+            await Task.CompletedTask;
             
             return new CloneResult();
         }
